Create structure for header and hero section
Lays out the html for the header, as well as creates the basic structure
for the hero section of the page. I created a div to contain the whole
hero section, a div for all the text content, and a div for the image.
Inside the text content are three divs for the heading, content and
button in the hero text.