Create structure of info section
Adds html to begin structuring the info section. Two main divs, a header
and a container for the cards section. Each card then is its own div,
containing a div for the image and a div for the text. In section will
be column, cards will be row, and each card will be column.