Finish buttons and begin styling main
Finishes css for styling of buttons and begins applying flexbox to
properly align content. Having some difficulties in the main section.
Need to figure out how to get the header to only expand to the size of
the hero section underneath it. Grouped both header and hero into a main
div for center alignment. May need to nest them one div further in order
to properly center content and allow header to expand to same size as
hero.