Format links and add padding to hero
Adds padding to hero to space it out from header and bottom of main.
Adds flex to ul to properly format links. Main sectoin is now properly
alligned, still need to add a little padding to header, and need to add
a placeholder to the image class to finish formatting main.