Add placeholder for image and finish styling main
Adds width constraints to image and hero text to style according to
assignment. In retrospect, I probably could have added a height and
width to the entire hero div instead of just hero text so that hero text
would be able to expand and contract depending on the size of the given
image. Also adds padding to top of header. Adds placeholder text to
image class.