Align testimonial section
This commit removes flex from the entire document, properly applying a
uniform height to each main div. This also properly aligns the
testimonial section, as well as bolding and enlarging the thor class
to properly match assignment. Width was forced on this section in order
to make it appear correctly like the assignment. In the future, maybe
theres a way to achieve this without a set value for width and height,
allowing each section to expand and contract in proportion depending on
screen size. For now however, that is beyond my current scope.