Style footer and remove padding from info
Removed padding from info to make size uniform with other elements.
Added padding as well as flex to footer and centered on page. At this
point, it appears that all requirements of the assignment are met. That
being said, would be worth revisiting in the future to mess with font
type as well as simplify/ better organize css. In the future, will make
note to better organize css from the get go.